refactor: extract envelope assembly out of __main__
The provider fetch, the concurrent map over the selection and the pick of the
`active` provider were inline in main(), which meant any second in-process
consumer would have to grow its own copy of them.

Move them to aiusage/envelope.py as build()/enabled(). No behaviour change: the
emitted envelope is byte-identical and the provider selection is the same
comprehension it was before.

The local `envelope` variable in the --normalize branch is renamed to `raw`,
matching normalize()'s own parameter name, so it no longer shadows the module.
